requirements.txt
# YOLO classifier and dependencies
ultralytics>=8.2.0
# Pillow for image IO
Pillow>=10.0.0
# Requests for URL fetch
requests>=2.31.0
# Optional: OpenCV headless often required indirectly
opencv-python-headless>=4.8.0

# API stack
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0

